Add lots of codegen infrastructure needed for normalized caching TODO: Clean up codegen structure to separate responsibilities

+++ artemis_codegen/src/utils.rs
@@ -1,10 +1,10 @@
 use std::num::Wrapping;
 
-pub fn hash(x: &&str) -> u32 {
+pub fn hash(x: &&str) -> u64 {
     let x = x.as_bytes();
-    let mut h = Wrapping(5381u32);
+    let mut h = Wrapping(5381u64);
     for i in 0..x.len() {
-        h = (h << 5) + h + Wrapping(x[i] as u32)
+        h = (h << 5) + h + Wrapping(x[i] as u64)
     }
 
     h.0
